setting cookies

+++ store/models.py
@@ -17,7 +17,8 @@ class Customer(models.Model):
 
 class Product(models.Model):
     name=models.CharField(max_length=200)
-    price=models.FloatField()
+    price=models.DecimalField(max_digits=7,decimal_places=2) 
+    #max price of an item can have 7 decimal places => max_price<=1000000.00
     digital=models.BooleanField(default=False,null=True,blank=True)
     image=models.ImageField(null=True,blank=True)
